collection/requirements.txt
pyzmq
pandas
#influxdb
git+git://github.com/influxdata/influxdb-python@master
